// src/app/e-wallet/models/api/cardState.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type CardState = 'PendingIssuing' | 'NotActivated' | 'Active' | 'Lost' | 'Stolen' | 'Destroyed' | 'DoNotHonor';

export const CardState = {
    PendingIssuing: 'PendingIssuing' as CardState,
    NotActivated: 'NotActivated' as CardState,
    Active: 'Active' as CardState,
    Lost: 'Lost' as CardState,
    Stolen: 'Stolen' as CardState,
    Destroyed: 'Destroyed' as CardState,
    DoNotHonor: 'DoNotHonor' as CardState
};